drop cross_engine_tests file, fold into sync handler tests
